Create sqraper.php
BETA version. I am still treading with caution and adjustments will surely need to be made. In particular, the code has not yet been confirmed to work with posts that have images and/or posts that reference other posts. This is because, although in theory it SHOULD work, Q has not made any drops with these elements in them as of the date of publishing this script.

I ask that for those using this script, when you improve, fix a bug or make a correction, please post an updated script here for others to take advantage of. I worked hard on it for the community so please be a team player (WWG1, WGA) and help the community back. Don't be "that guy".

The script is not intended to be run as a web page. It is intended to be run as a script, such as in Windows "PHP.exe sqraper.php". PHP must be installed on your computer. At the date of publishing, I have only tested it on a Windows installation. I was not planning on releasing it yet, but due to some other scrapers lagging in picking up posts, I decided to roll it out to a few early to help others overcome the delays.